h264: Populate nal_ref_idc
Although VAAPI doesn't really provide such info, we can fake it from
reference_pic_flag info. It seems that no decoder, SW or HW, really
cares what is the exact value of this parametes, just if it is != 0.

Fake it by assigning value 3 if reference flag is set to true.

Signed-off-by: Jernej Skrabec <jernej.skrabec@siol.net>